Update and rename fileminio.py to MinioHandler.py
